--- a/dashboard/package.json
+++ b/dashboard/package.json
@@ -22,7 +22,7 @@
     "babel-preset-react-hmre": "~1.1.1",
     "bootstrap-loader": "~1.0.10",
     "bootstrap-sass": "~3.3.7",
-    "css-loader": "~0.23.1",
+    "css-loader": "0.14.5",
     "fetch-ponyfill": "^3.0.2",
     "file-loader": "~0.9.0",
     "hjs-webpack": "~8.3.0",