--- a/dashboard/src/config/webpack.app.js
+++ b/dashboard/src/config/webpack.app.js
@@ -1,4 +1,5 @@
 /*eslint-env node*/
+
 // TODO: this should be broken up into `dev` and `prod`
 // configuration variants