--- a/dashboard/src/config/webpack.base.js
+++ b/dashboard/src/config/webpack.base.js
@@ -1,3 +1,5 @@
+// Original from https://github.com/mxstbr/react-boilerplate/
+
 /*eslint-env node*/
 
 /**