--- a/dashboard/src/features/shared/actions/create.js
+++ b/dashboard/src/features/shared/actions/create.js
@@ -1,4 +1,4 @@
-import chain from 'chain'
+import chain from '_chain'
 import { context } from 'utility/environment'
 import { parseNonblankJSON } from 'utility/string'
 import actionCreator from './actionCreator'