--- a/dashboard/src/utility/environment.js
+++ b/dashboard/src/utility/environment.js
@@ -1,6 +1,6 @@
 /* global process */
 
-import chain from 'chain'
+import chain from '_chain'
 import { store } from 'app'
 
 import { useRouterHistory } from 'react-router'