Reduces Dashboard development startup time from 15s to 3s. 
Reduces Dashboard development rebuild time from 4s to 500ms-1s.

This PR packages unchanging dependencies in a webpack DLL, allowing
development server to avoid reparsing all dependencies individually
with each rebuild.